Completed XML reader.
