+++ BrickBreaker/Screens/GameScreen.cs
@@ -1,7 +1,7 @@
 ﻿/*  Created by: 
  *  Project: Brick Breaker
  *  Date: 
- */ 
+ */
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
@@ -12,6 +12,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 using System.Media;
+using System.Xml;
 
 namespace BrickBreaker
 {
@@ -24,6 +25,7 @@ namespace BrickBreaker
 
         // Game values
         int lives;
+        int x, y, width, height, id;
 
         // Paddle and Ball objects
         Paddle paddle;
@@ -72,24 +74,56 @@ namespace BrickBreaker
             int ballSize = 20;
             ball = new Ball(ballX, ballY, xSpeed, ySpeed, ballSize);
 
-            #region Creates blocks for generic level. Need to replace with code that loads levels.
-            
-            //TODO - replace all the code in this region eventually with code that loads levels from xml files
-            
-            blocks.Clear();
-            int x = 10;
+            LevelReader();
 
-            while (blocks.Count < 12)
-            {
-                x += 57;
-                Block b1 = new Block(x, 10, 1, Color.White);
-                blocks.Add(b1);
-            }
+            #region Creates blocks for generic level. Need to replace with code that loads levels.
+
+            //TODO - replace all the code in this region eventually with code that loads levels from xml files
+
+            //blocks.Clear();
+            //int x = 10;
+
+            //while (blocks.Count < 12)
+            //{
+            //    x += 57;
+            //    Block b1 = new Block(x, 10, 1, Color.White);
+            //    blocks.Add(b1);
+            //}
 
             #endregion
 
             // start the game engine loop
             gameTimer.Enabled = true;
+        }
+
+        public void LevelReader()
+        {
+            XmlReader reader = XmlReader.Create("Resources/TempXML.xml");
+
+            while (reader.Read())
+            {
+                if (reader.NodeType == XmlNodeType.Text)
+                {
+                    x = Convert.ToInt32(reader.ReadString());
+
+                    reader.ReadToNextSibling("y");
+                    y = Convert.ToInt32(reader.ReadString());
+
+                    reader.ReadToNextSibling("width");
+                    width = Convert.ToInt32(reader.ReadString());
+
+                    reader.ReadToNextSibling("height");
+                    height = Convert.ToInt32(reader.ReadString());
+
+                    reader.ReadToNextSibling("id");
+                    id = Convert.ToInt32(reader.ReadString());
+
+                    Block newBlock = new Block(x, y, width, height, id);
+                    blocks.Add(newBlock);
+                }
+
+            }
+
         }
 
         private void GameScreen_PreviewKeyDown(object sender, PreviewKeyDownEventArgs e)
@@ -199,7 +233,7 @@ namespace BrickBreaker
             // Goes to the game over screen
             Form form = this.FindForm();
             MenuScreen ps = new MenuScreen();
-            
+
             ps.Location = new Point((form.Width - ps.Width) / 2, (form.Height - ps.Height) / 2);
 
             form.Controls.Add(ps);
