--- GenXML.xml
@@ -1,18 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<level>
-	<brick>
-		<x>201</x>
-		<y>32</y>
-		<width>50</width>
-		<height>25</height>
-		<id>5</id>
-	</brick>
-	
-	<brick>
-		<x>312</x>
-		<y>122</y>
-		<width>147</width>
-		<height>116</height>
-		<id>1</id>
-	</brick>
-</level>
